+++ app/src/main/java/com/example/deliveryapp/ui/components/menu/Menu.kt
@@ -1,4 +1,4 @@
-package com.example.deliveryapp.ui.menu
+package com.example.deliveryapp.ui.components.menu
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
