+++ app/src/main/java/com/example/deliveryapp/ui/components/menu/MenuItem.kt
@@ -1,4 +1,4 @@
-package com.example.deliveryapp.ui.menu
+package com.example.deliveryapp.ui.components.menu
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -6,11 +6,7 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
