+++ app/src/main/java/com/example/deliveryapp/ui/components/menu/MenuSection.kt
@@ -1,10 +1,8 @@
-package com.example.deliveryapp.ui.menu
+package com.example.deliveryapp.ui.components.menu
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
